worked on programming protocol implementation. not 100% done. now needs arduino counterpart for first testing and finetuning...

+++ src/main/java/de/konnekting/deviceconfig/utils/Helper.java
@@ -269,17 +269,17 @@ public class Helper {
         return createTempFile.getName();
     }
     
-    public static byte getHI(short v) {
+    public static byte getHI(int v) {
         return (byte)((v >>> 8) & 0xFF);
     }
     
-    public static byte getLO(short v) {
+    public static byte getLO(int v) {
         return (byte)((v >>> 0) & 0xFF);
     }
     
-    public static short getFromHILO(byte hi, byte lo) {
+    public static int getFromHILO(byte hi, byte lo) {
         int ch1 = hi;
         int ch2 = lo;
-        return (short) ((ch1 << 8) + ((ch2 << 0) & 0xFF));
+        return (int) ((ch1 << 8) + ((ch2 << 0) & 0xFF));
     }
 }
